Dashboard: Cập nhật lại Statistic
